requirements.txt
setuptools
wheel
django
django-markdown
oauth2-provider
django-cors-headers
django-oauth-toolkit
django-rest-framework
djangorestframework-xml
djangorestframework-constant-field
pillow
pycrypto
